Add migrations and refactor database schema
- Created a new folder for Migrations in the DataBaseService project.
- Added new entities and configurations in the database context, including HistoryOfIrrigation, HistoryOfTankWaterLevel, and IrrigationSchedule.
- Updated the Tank model to include a Name property and changed data types for TankId and SprinklerId in HistoryOfIrrigation.
- Introduced a new migration class `ReworkDb` to define the database schema for various models.
- Wrapped GetAllAsync method in a try-catch block for error handling and logging.
- Updated Dockerfile and connection string logic for environment differentiation.
- Refactored BaseSpecification class name for consistency.
- Modified XAML files to use new command bindings and updated model snapshot files.

+++ Model/HistoryOfIrrigation.cs
@@ -8,9 +8,9 @@ namespace Model
         public TimeSpan StartTime { get; private set; }
         public TimeSpan EndTime { get; private set; }
         public TimeSpan Duration { get; private set; }
-        public int TankId { get; private set; }
+        public long TankId { get; private set; }
         public Tank Tank { get; private set; }
-        public int SprinklerId { get; private set; }
+        public long SprinklerId { get; private set; }
         public Sprinkler Sprinkler { get; private set; }
         public double AmountOfWaterUsed { get; private set; }
         public double WaterLevelBefore { get; private set; }
